# Optimize Estrin evaluation by stopping the recursion at 0 and 1 and doing more inlining