Now uniformly faster than Horner:
```
.\Release\x64\benchmarks.exe --benchmark_filter=Poly
Run on (4 X 3310 MHz CPU s)
01/14/18 11:00:52
---------------------------------------------------------------------------------------------------------------
Benchmark                                                                        Time           CPU Iterations
---------------------------------------------------------------------------------------------------------------
BM_EvaluatePolynomialInMonomialBasisDouble<EstrinEvaluator>/4                 2042 ns       2079 ns     345165
BM_EvaluatePolynomialInMonomialBasisDouble<EstrinEvaluator>/8                 3542 ns       3438 ns     195094
BM_EvaluatePolynomialInMonomialBasisDouble<EstrinEvaluator>/12                5099 ns       5006 ns     112179
BM_EvaluatePolynomialInMonomialBasisDouble<EstrinEvaluator>/16                7192 ns       7127 ns      89743
BM_EvaluatePolynomialInMonomialBasisQuantity<EstrinEvaluator>/4               1977 ns       1989 ns     345165
BM_EvaluatePolynomialInMonomialBasisQuantity<EstrinEvaluator>/8               3545 ns       3518 ns     195094
BM_EvaluatePolynomialInMonomialBasisQuantity<EstrinEvaluator>/12              4885 ns       4958 ns     144747
BM_EvaluatePolynomialInMonomialBasisQuantity<EstrinEvaluator>/16              7289 ns       7301 ns      89743
BM_EvaluatePolynomialInMonomialBasisR3ElementDouble<EstrinEvaluator>/4        5046 ns       5148 ns     100000
BM_EvaluatePolynomialInMonomialBasisR3ElementDouble<EstrinEvaluator>/8        9640 ns       9595 ns      74786
BM_EvaluatePolynomialInMonomialBasisR3ElementDouble<EstrinEvaluator>/12      15945 ns      15992 ns      44872
BM_EvaluatePolynomialInMonomialBasisR3ElementDouble<EstrinEvaluator>/16      21250 ns      20929 ns      32051
BM_EvaluatePolynomialInMonomialBasisVectorDouble<EstrinEvaluator>/4           5211 ns       5148 ns     100000
BM_EvaluatePolynomialInMonomialBasisVectorDouble<EstrinEvaluator>/8          10012 ns       9978 ns      64102
BM_EvaluatePolynomialInMonomialBasisVectorDouble<EstrinEvaluator>/12         14554 ns      14706 ns      49857
BM_EvaluatePolynomialInMonomialBasisVectorDouble<EstrinEvaluator>/16         19872 ns      19886 ns      34517
BM_EvaluatePolynomialInMonomialBasisDisplacement<EstrinEvaluator>/4           5075 ns       5148 ns     100000
BM_EvaluatePolynomialInMonomialBasisDisplacement<EstrinEvaluator>/8          11116 ns      11195 ns      64102
BM_EvaluatePolynomialInMonomialBasisDisplacement<EstrinEvaluator>/12         15001 ns      14949 ns      44872
BM_EvaluatePolynomialInMonomialBasisDisplacement<EstrinEvaluator>/16         18961 ns      19191 ns      37393
BM_EvaluatePolynomialInMonomialBasisDouble<HornerEvaluator>/4                 1922 ns       1898 ns     345165
BM_EvaluatePolynomialInMonomialBasisDouble<HornerEvaluator>/8                 3933 ns       3911 ns     179486
BM_EvaluatePolynomialInMonomialBasisDouble<HornerEvaluator>/12                7717 ns       7649 ns      89743
BM_EvaluatePolynomialInMonomialBasisDouble<HornerEvaluator>/16               11664 ns      11682 ns      56089
BM_EvaluatePolynomialInMonomialBasisQuantity<HornerEvaluator>/4               1922 ns       1898 ns     345165
BM_EvaluatePolynomialInMonomialBasisQuantity<HornerEvaluator>/8               3963 ns       3911 ns     179486
BM_EvaluatePolynomialInMonomialBasisQuantity<HornerEvaluator>/12              7677 ns       7649 ns      89743
BM_EvaluatePolynomialInMonomialBasisQuantity<HornerEvaluator>/16             11473 ns      11403 ns      56089
BM_EvaluatePolynomialInMonomialBasisR3ElementDouble<HornerEvaluator>/4        5107 ns       5148 ns     100000
BM_EvaluatePolynomialInMonomialBasisR3ElementDouble<HornerEvaluator>/8       10578 ns      10465 ns      64102
BM_EvaluatePolynomialInMonomialBasisR3ElementDouble<HornerEvaluator>/12      16385 ns      16340 ns      44872
BM_EvaluatePolynomialInMonomialBasisR3ElementDouble<HornerEvaluator>/16      24311 ns      23919 ns      28045
BM_EvaluatePolynomialInMonomialBasisVectorDouble<HornerEvaluator>/4           5208 ns       5304 ns     100000
BM_EvaluatePolynomialInMonomialBasisVectorDouble<HornerEvaluator>/8          10948 ns      10569 ns      56089
BM_EvaluatePolynomialInMonomialBasisVectorDouble<HornerEvaluator>/12         18991 ns      18530 ns      34517
BM_EvaluatePolynomialInMonomialBasisVectorDouble<HornerEvaluator>/16         24563 ns      24475 ns      28045
BM_EvaluatePolynomialInMonomialBasisDisplacement<HornerEvaluator>/4           5223 ns       5304 ns     100000
BM_EvaluatePolynomialInMonomialBasisDisplacement<HornerEvaluator>/8          10774 ns      10951 ns      64102
BM_EvaluatePolynomialInMonomialBasisDisplacement<HornerEvaluator>/12         19440 ns      19434 ns      34517
BM_EvaluatePolynomialInMonomialBasisDisplacement<HornerEvaluator>/16         24385 ns      24510 ns      29914
```